rest/requirements.txt
Django==2.1.1
djangorestframework==3.8.2
jsonpickle==1.0
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
